safari-600.7-branch/LayoutTests/fast/workers/storage/empty-statement-sync.html

<html>
<head>
<script src="resources/test-inputs-common.js"></script>
</head>

<body onload="runTest('resources/empty-statement-sync.js')">
This test tries to execute an empty statement.
<pre id="console">
</pre>
</body>
</html>
