safari-600.7-branch/LayoutTests/fast/workers/storage/empty-statement-sync-expected.txt

This test tries to execute an empty statement.
PASS: result is null
PASS
